pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.52.2",
    "apscheduler>=3.11.0",
    "dropbox>=12.0.2",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "google-generativeai>=0.8.5",
    "gunicorn>=23.0.0",
    "matplotlib>=3.10.3",
    "notion-client>=2.3.0",
    "numpy>=2.2.6",
    "openai>=1.84.0",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
]
